requirements.txt
ldap3
requests
pyyaml
